Disallow leading dash and underscore in project name
Dashes and underscores are currently permitted in project name.
Leading dashes and underscores are not legal in docker
volume names.
This change removes leading dashes and underscores from project
names so that the subsequent volume names are legal.
This is a minor breaking change, as existing stacks may have
leading dashes and underscores. However, if they do, then
docker volumes generated by this stack will be illegal and
therefore it is unlikely that these are common.